+++ lib/src/graphql/graphql.dart
@@ -1,3 +1,4 @@
 export 'fresh/fresh.dart';
+export 'model/cursor_page.dart';
 export 'repo/repo.dart';
 export 'utils/utils.dart';